backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
httpx==0.27.0
pymupdf==1.24.2
python-docx==1.1.2
openpyxl==3.1.2
arq==0.25.0
redis==5.0.4
modal==0.62.0
unsloth==2024.5
python-dotenv==1.0.1